Major refactor of the network simulation
I *think* this should be a lot more efficient and also more flexible because it
uses flexible classes underneath instead of a lot of magic strings and
dictionaries. It also makes the logic more self contained and should allow
better separation in the future.
So overall I think it's an improvement and wasn't a waste of time :)